ptest: fix test build with latest qml tools
Qml tools are now using --bare option, which limits where the tools
tries to find the qml modules. For the test builds, we again need to
use QT_ADDITIONAL_PACKAGES_PREFIX_PATH to tell the tools where to
find qml files for the Qt module we are currently building.

Task-number: QTBUG-106365
Pick-to: dev 6.4
Change-Id: I6dfcd32f405475de0f58448cbaa7a19e26cbcfe8
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>