qtwebengine: Pass OE specific pkg-config-native for host pkg-config
chromium has a way to use right pkg-config for in-tree hosttools which
are built and used during cross build of qtwebengine, since OE
pkg-config environment is set for detecting .pc files from sysroot,it
ends up reading wrong pkg-config files for host pieces and adds
-L<sysroot>/usr/lib to linker cmdline for hosttools too, this goes by
fine with gnu BFD linker since it ignores linker script files like
<sysroot>/usr/lib/libc.so, but when we use LLD linker, it ends up in
error, because it tries to process the content of libc.so which have
target absolute paths for ldso etc. /usr/lib/ld-linux-aarch64.so.1
LLD tries to open these files and ofcouse can't find them because
they are only found in target sysroot.

The underlying problem is that we need to be explicit about pkg-config
which is  thusly implemented by this patch. This fixes build with
clang+lld and link times for qtwebegine is reduced by 26% in my case.

Pick-to: 6.7
Change-Id: Id091200444eb97d4b44d984888a8aa16b75ff9d7
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>