Remove old patches from qtwebengine
The workaround for long rsp file is an old patch
from qt5 times, which should not be needed in qt6
as used toolchains are in-source, therefore created
toolchain labels and later rsp files should not embedded
absolute paths, but simply use 'build/toolchain/yocto'
as this is a location of yocto toochain with-in source
tree.

The second patch is fixed in 112-based in 29354c7c7de.

Task-number: QTBUG-59769
Change-Id: I94a610c4562463f5de3f8bbb29915e05cf40e640
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>
(cherry picked from commit bdbb28f3624c2df9b9dd0521ac2924378ccb8916)