ptest: better handling for test list
Use ctest to find all tests to run as not all of them start with
a tst_ prefix. This is done by parsing the json output of ctest.
Add all test sources to the package as there are data files in
some tests that are needed.

Remove qmake support, since it's not supported when building
Qt module.

QtDeviceUtilities had includes in wrong order, so ptest runner was
never added to the package.

Task-number: QTBUG-118680
Pick-to: 6.7 6.6
Change-Id: I6397cb1fd331941a3425bf627676b6387ab8bb8d
Reviewed-by: Ari Parkkila <ari.parkkila@qt.io>