qtwebengine: add PACKAGECONFIG for x11 dependencies
webengine-ozone-x11 feature (Support GLX on qpa-xcb) has dependencies
for x11 libraries and libgl. Add all of them into new PACKAGECONFIG
that's enabled if x11 and opengl are in DISTRO_FEATURES.

libxdamage is added as new dependency as it's needed when both
ozone-x11 and webrtc are enabled.

libgl is added to make sure it's available regardless of how qtbase
has been configured (e.g. with gles2) and dependencies are not met
through it.

Pick-to: 6.7
Change-Id: I6708d3440b7f1ef5b3b8dcc3a1563e6e5fc48dc0
Reviewed-by: Ari Parkkila <ari.parkkila@qt.io>
Reviewed-by: Khem Raj <raj.khem@gmail.com>
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>