qtwebengine: set NINJAFLAGS variable to $PARALLEL_MAKE
According to documentations to build QtWebEngine, parallel option
set by CMake is not taken into account during build:

"Note: The --parallel CMake option is not used for QtWebEngine build.
It will not have any observable effect because of the recursive ninja call.
Use NINJAFLAGS environment variable instead."

This change is restoring PARALLEL_MAKE variable effect to reduce the number of jobs
to build this recipe. This is important due to high resources consumed by
this build.

Pick-to: 6.11 6.8
Change-Id: I08a0df8c16d698348a15fe5416c1d84fd423bc3f
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>